-> New Euler: "Unglue Face Region Make Vert"
Added a new euler that will take as an argument a face
that is part of a disk and a vertex in the face. The euler
will then 'unglue' this region. An example of this would
be two cones joined at the tip...

Note that this code is untested and probably will have
bugs so shouldnt be trusted yet...