-> Added new restrction to Join Face Kill Edge Euler
Join Face Kill Edge now checks to make sure it wont create
a face where the same vertex appears twice in the loop cycle.

Note to Joe:

This is what we talked about on IRC a while back. It
seems to work from here, but you should probably give it a
really good test in the vert dissolve code.