Added a matrix slot type for bmops.  Coded a simple transform bmop, that just multiplies input verts with a matrix.  Also made a derivative translate bmop.
BMO_CallOpf now has a %s format code, which is used to
copy data from another slot.

Also cleaned the extrude code up some more, and restored extrude-repeat
(which is bound to ctrl-alt-4), though this doesn't work right yet
(the view matrix it uses is incorrect, or something like that).