indentation error with some operators (strip the descriptions) if 0'd the exec() workaround for running python scripts incase windows devs want to test.
theeth said the problem is when you compile a debug blender against a non debug python (or the ther way I assume), you can get a FILE struct mismatch.

A quick way to test this on windows is to run this from the command line.
 blender -P somescript.py
 
Somescript.py can be anything