Feature fix
When using Material Nodes, there was no indication in Material
Property window which node was active. The context/channel widget
now shows this.

Better would be to be able to browse nodes in this list, and
to have option to preview only this material (not node tree
result). Enough todo for future :)