Button for adding shape keys now creates shapes from base mesh instead of from the current shape mix.  The old behavior is still accessable from the menu as "New Shape From Mix".
Checked with Sergey and Bassam that this is a good change.  New users
expect the add shape button to simply add a new blank shape, and get
confused when that is not the case.  It is also really easy to
accidentally have other shape information in a new shape when the
"from mix" behavior is default.