fixed crash on loop cut of wire edge
Part of fix was to make edgering walker work when initial
edge is a wire edge (has no faces attached): it will just
walk that one edge and quit now.  I checked other uses
of the edgering walker and this won't hurt them.  Had
to add a new wireedge variable to the edge ring walker state.