Fixed a crash due to invalid pointers to temporary scenes for stroke rendering.
Freestyle generates a temporary scene containing a number of stroke meshes to
achieve stroke rendering using the Blender internal renderer.  This temporary
scene may appear in the outliner and undergo user interaction when the stroke
rendering takes a long time.  The crash was due to the deletion of the scene
without taking care of references to the scene from screens, one of which was
the outliner.