Fix for an inappropriate removal of singular points in stroke creation.
The previous stroke creation procedure was trying to clean stroke topology
by removing overlapping stroke vertices in the same 2D location.  The idea
was to avoid having to address this kind of singularity during subsequent
stroke shading.  In-depth analyses revealed, however, that this was a wrong
way to ensure clean stroke topology, since just deleting overlapping vertices
may break the continuity of the underlying series of FEdges on top of which
the stroke has been built.  Such a break of linked FEdges was a major cause
of frequent failure in CurvePoint::getFEdge().

The present commit aims to address the singularity issue by adding small
offsets to the 2D location of overlapping vertices and making them
non-overlapping to each other.  Since the offsets only result in sub-pixel
differences, the impact on visual outcomes is expected to be negligible.