ACES ODT tonecurve style cleanup and small tweaks
Convert color into ACES color space from rec709 (which is quite the same
as blender internal) before running tonecurve and convert color space back
after tonecurve.

This makes tonecurve be applying in actual space it was intended to and it
makes color range smoother a bit.

Would be nice to convert this into OCIO view tho.