bge bugfix: [#18967] alpha value is never available in the color buffer by Campbell Barton (X11) and me (Win32) This is and old patch (June, 2009). The reason it never made into trunk it's that we (me at least) thought that it should be optional, as a ui or command-line option.
However, it seems that OSX always have the OpenGL context with alpha enabled. So I think it's consistent to
have the other OSs to follow.

The main usage of this is the BGE (or more specifically people using BGE for TV broadcasting) but I think
pydevs can have their share of fun with it in Blender as well.