Statistics calculation for threaded update
This commit adds per-thread statics for object
update threads which would give you information
about:

- How much objects each thread handled
- How much overall time thread spend on running
  object_handle_update.
- How long each of object_handle_update took.

Enabled by ./blender -d

The code is surrounded by ifdef, so shall be
not a problem to drop the code when we don't
need it anymore.

Also added special value for rt (debug_value)
of 13666 which switches scene update to a
single thread. Useful for benchmarking.