Fix T38651: Compositor Normal Node Sphere unchangable when Reset All To Default Values is used on it.
The default for the socket value was always 0. Now the default value for
node socket default_value properties is chosen based on the socket
subtype to give some more sensible values.

NOTE: this may still be different from the values specified in the
socket templates! It is not possible to directly use these template
values, because all sockets share the same RNA types. At least it should
work reasonably well in the majority of cases now.