NLA Editor: Make Single User for NLA Strips
This operator is used to make sure that if/when you have multiple strips
using the same action, if you select these and run this operator, each
strip will be given its own copy of the action. This is useful if you
decide later that you want to start using an existing action as a base.

NOTE: This does not recursively go inside meta's, so care is still advised
in that case.