Fix issue discovered in soc-paint-2013 branch.
Weights do not need perspective correction since they are calculated on
flat triangle coordinates. Only pixel coordinates, reprojected from the
weights need this.

This was already fixed for clone brushes but in
soc-2013-paint where we have stencil preview, it is apparent that the
stencil suffered from the perspective distortion issue too.