Revert "fix T40375 Glossy shader bakes different than render"
This reverts commit 12abe94de827d9ae9c0dd6cc49bc6c3e377842ad.

After a long discussion in the bug tracker we decided baking should use
the faces normals for glossy (and combined). This is what Blender
Internal is doing, and one of the more predictable way of yielding
predictable results.

That also means the result will not match the render perfectly, but this
is preferrable over the alternatives at hand.

Conflicts:
	intern/cycles/kernel/kernel_bake.h