Keymaps: make click event use position on button press rather than release.
This may improve reliability with left click select and pen input, assuming
that the place where the pen first touched the surface is closer to the
intended location than where it was released from the surface.

I'm not sure if this will make a significant difference in practice, but it
seems worth a try.