Fix T58391: correctly handle Curve Shape Key layout in Python API.
The layout changed when the radius property was added to shape
keys in 2.8, but the RNA code wasn't updated.

Also, even before that, the code didn't do anything to correctly
handle mixing sub-curves of different type (nurbs vs bezier) in
the same Curve object. Now that case is handled correctly but not
very efficiently by allocating a mapping table when necessary. To
recover some performance, a custom index lookup function is added.