Fix access to current preferences when version patching read preferences
The version patch for 0cf9794c7ef0 was checking and setting a data name
using the macros for translation. These would access the current
preferences which can mismatch the ones currently being version patched.
See discussion in 0cf9794c7ef0 for details.

Don't handle translation in this version patch, which is more of a
"nice-to-have" version patch, no functionality depends on it.