Fix: Hide "Squeeze Value" node from node search
Based on discussion in D10891, this node isn't meant to be exposed and
may be removed in the future. The fact that it was exposed in search
menus was a mistake from the implementation of link-drag-search
and bdb57541475f20ccc4.

I explicitly removed the link drag search implementation, and added
(Legacy) to the node name which hides it from the add node search.