Revert "Fix: Knife tool does not interpolate vertex customdata in interior cuts"
This reverts commit 129f79debe9b693251db8728f7619914a1aefd30.

That commit changed the behavior of how booleans handled vertex weights
as well and made the CubeMaskFirst test modifier test fail.

Not entirely clear to me what the desired behavior would be (current
situation is "wrong" in certain situations as well I think), but until
this is further discussed with the #modeling-module , I think reverting
is the better choice.