Fix #108211: Cycles: Correctly split Glass BSDF contributions
So far, each closure in Cycles was either diffuse OR glossy OR
transmissive, and its color and contributions were assigned
to the corresponding direct/indirect/color passes.

However, since Glass is a single closure now, that is no longer enough,
since glass has both a glossy and a transmissive component.

Therefore, this commit adds support for splitting contributions from
the Glass closure between the two types.
For 4.0, we might want to also use this for Principled Hair since it
also technically has both types, but that would be a change from
the existing result so it's not part of 3.6 yet.