Fix #108704: scrubbing timeline does not to proper sampling like normal frame change
This was broken in 037b3f87bd58e003d8fe8ca725dfa59969657e6c.
This fix brings the original problem back a bit in that there will be an additional
frame update now, but really only a redraw should be necessary without a
depsgraph update.

The depsgraph update is caused by the `NC_SCENE | ND_FRAME` notifier,
which is checked for in `wm_event_do_notifiers`. Changing that in more depth
is a bit risky for 3.6 now unfortunately.