Geometry Nodes: Simplify code in duplicate elements face mode
This node splits all faces to duplicate each time. Corner edges and
vertices are just ordered sequences of indices. Most likely, due to
historical reasons, this was overlooked and as a result the code became
so complex with potential simplicity. The problem was found in a
benchmark of #109174. The code in the lambda could not be optimized.

Timings improved by about 30%, from 154 to 120 ms.

Pull Request: https://projects.blender.org/blender/blender/pulls/109222