Fix EEVEE: Broken Normals and Transparency
This was caused by 89218b66c2192f6d517776211855f0306c3d5f57
The removal of the transmission roughness socket messed
with the hardcoded socket indices used to for auto socket
values and optimization detection.

Use defines inbetween sockets declarations to make it more
obvious that it they needs to be updated on socket changes.
This also makes the code a bit more meaningful.