Fix #109403: GP doesn't render selection outline
The issue was that there was a change in the shader that expected
a specific flag to be set. The current grease pencil code did not
set these flags.

This fix separates the old and new shader code so that
the old way of rendering the strokes remains untouched.

Pull Request: https://projects.blender.org/blender/blender/pulls/111227