Animation: Graph Editor optimize handle drawing
The drawing code of the keyframe handles had the same issue
as the curve drawing itself, namely that it drew everything even things out of view.

This patch uses the same logic introduced in #110301 to get the
bounding indices of the curve in the viewport and only draw those.

This does NOT improve the performance when fully zoomed out on large data.

Test setup: 6000f of dense key data on 62bones
I ran my test on a zoom level that had the Graph Editor display about 500 frames.
The numbers are per curve and ONLY the handle drawing functions.

| Before | After |
| - | - |
| ~300μs | ~53μs |

Pull Request: https://projects.blender.org/blender/blender/pulls/112126