Fix #112247: Node group socket declarations using only 'NONE' subtype
The declaration code for node groups was relying on `bNodeSocketType` to
provide the subtype of a socket. This worked before node panels
(#111348) since the interface sockets had a fully refined typeinfo.

Now the interface sockets use only the base typeinfo and the socket
subtype is stored in the interface data itself.

Pull Request: https://projects.blender.org/blender/blender/pulls/112286