Fix: Dyntopo sculpt on plane destroys the mesh
The issue is probably deeper and affects more cases. The plane with an
edge loop in the center added was the case which shown the error.

Caused by a mistake in the refactor in 062169698a7a which left an array
size uninitialized before filling it in.

Pull Request: https://projects.blender.org/blender/blender/pulls/113972