Fix: GPv3: Stutter while drawing
While drawing, the line beeing drawn would seemingly only update once
in a while. This was because there was a bug in the interpolation
function that would write the value being interpolated from
directly to the destination as the first value.
In our case, we only wanted to write new values (so
we need to exclude the first one).

Making sure the interpolation always excludes the value
interpolated from, fixes the drawing stutter issue.