Fix #115181: RMB Select sometimes to enter pose mode when selected
Pass-through selection allows select-drag not to activate an already
selected object - so an existing selection can be dragged without
changing the active/selection (see #96544) & [0],

This caused entering pose-mode to fail when selecting an armature
with "Lock Object Modes" disabled.

There is a logical conflict where pass-through is intended to keep the
current selection & active, which doesn't make sense when selecting the
armature should enter pose-mode (details in code-comments).

Resolve by disabling pass-through when the select action changes
the objects mode.

[0]: 618f39fca23ca7734a09a8f08efced8379efd2b3