Fix: Mesh draw normals extraction ignores sharp_edge with no sharp_face
See c4446d79245125f3484041d06510ff75fbedc509

When the "fully flat" state comes from "sharp_edge" and "sharp_face"
doesn't exist, we need to check for that for every face when extracting
normals. Eventually these loops should be unrolled so we don't have a
function call per face. That would remove the cost of this check.