Assets: Speed up iterating over asset libraries
Essentially this avoids the file browser entry cache mechanism for when
we only want the list of assets, and don't need the preview image. The
speedup should become bigger the bigger the asset library is, see #120494.

When iterating over assets, we would do that through the `AssetHandle`
type, which is just a wrapper around a cached file in the file browser
backend (for historic reasons). So first we'd have to ensure this entry
is actually cached, which involved some lookups and a bunch of
allocations, which would easily add up.

In fact, for as long as we don't need the asset-handle (usually to get
the preview), we have the needed data easily available already, without
requiring a cached file in the file browser backend.

See the pull request for benchmark results.

Pull Request: https://projects.blender.org/blender/blender/pulls/120699