EEVEE-Next: UI: Change default filtering radius to 1px
The resonning is that we expect the default to be close
to realistic rendering. The filtering is only here to
hide aliasing in this case and shouldn't do more
blurring. This is specially disturbing when rendering at
low resolution since the filter is resolution dependant.

Older files are also expecting to use a 1px filter
for compatibility.