GPv3: Draw Tool: Angle option
This implements the angle and angle factor draw tool settings in the `Advanced` panel.

The behavior is changed slightly to GPv2. Instead of calculating the radius of the point
based on an interpolated value between the adjusted radius and the previous point,
in GPv3 the radius is adjusted by a "smoothed angle factor". This factor is the angle factor
based on the previous point, but smoothed over time (by a constant amount).

Pull Request: https://projects.blender.org/blender/blender/pulls/122488