Fix #123328: Onion skinning causes change in opacity for current frame
Toggling the layer onion skinning toggle would affect the opacity
of strokes on the current frame.

This was caused by the `get_visible_frames_for_layer` function. It computes
a `frame_id` for all the keyframes that are not the currenty visible
keyframe. To skip over the currently visible keyframe on the layer,
it just compared the start frame of the keyframe with the current
scene frame. This only works if the current frame is over the keyframe,
but not for any frames after (even if the same keyframe is still visible).

The fix computes the start frame of the keyframe under the current frame
first, and then uses that to compare to the start frame during the iteration
over all keyframes. This makes sure that we skip over the currently visible
keyframe in that layer.

Pull Request: https://projects.blender.org/blender/blender/pulls/123358