Sculpt: Move some mesh geometry storage out of per-node undo data
This storage of the entire mesh for topology-changing operations doesn't
need to be stored in every single undo node. Move it to the data for the
undo step instead. This decreases the size of undo nodes from 3880 to
1816 bytes. That saves about 4MB of memory on a single stroke affecting
most of a 6 million vertex mesh.

I didn't change anything BMesh related here because it's trickier to get
right and not quite as encapsulated. Moving all BMesh undo data out
of `undo::Node` would be a good step though, because only one undo
node is used anyway.