Sculpt: Initial data-oriented refactor for clay brush
Part of #118145.

Performance after this change results in an 119% speedup. This is
measured on the fucntion `stroke_update_step` on a mesh with 24k verts
across 5k instances of the function being timed. In absolute values,
this represents a 0.06ms speedup, from 0.38ms to 0.32ms.

Pull Request: https://projects.blender.org/blender/blender/pulls/123751