Fix #120465: Crash using the snap tool with Wayland & Libdecor
The snapping checked use the wmWindowManager::winactive without a null
check. This is documented to be NULL at times.

Resolve by using the context's window which makes sense as all other
screen variables are already using the context.

Libdecor is also missing window activation on startup,
this will be fixed separately.