Fix #125509: Reordering of channels doesn't work in Graph Editor
The reordering of channels didn't work in the following case:
* in the Graph Editor
* Channels outside of a FCurve group

The reason for that was that any channels outside a group get added to a temp group
(sorting somehow depends on groups) and that was missing the `AGRP_EXPANDED_G`
flag. The reason this was only broken in the Graph Editor was that this flag is only used
for the graph editor. The `EXPANDED_AGRP` macro has a special case for that
(used towards the end of `rearrange_action_channels`).

Pull Request: https://projects.blender.org/blender/blender/pulls/125673