UI: Detailed Large File & Folder Icons
Changes to the icons used for document and folder when in thumbnail
view in File Browser. A bit more detail, less chunky, more in tune
with OS versions. Better match for the recent increased fidelity of
the "type" icons that are shown in their centers.

Pull Request: https://projects.blender.org/blender/blender/pulls/126760