Fix key_command icon
This error was introduced on 7dc3e8e04d97dd80d12f491e9277162c0289aa23
by the automatic script used (after my original PR code - !127194).

The SVG had the same attributes added twice.

The issue was technically introduced on
ffa802da7ac09273a34903438550e0966c4063dc (which is not in main but part
of the code which was merged).