Fix #99157: Allow Shortcut on Auto Keyframe
If toggling the status of the Auto Keyframe button with a custom
keyboard shortcut it does not update and display the current status of
the button - it looks like nothing has changed.  This PR just updates
the button correctly when used this way. Just the addition of a
RNA_def_property_update.

Pull Request: https://projects.blender.org/blender/blender/pulls/131406