Fix #131989: Mesh paint mode selection does not work
Caused by 839108f62376f10231f44087a9844c30d9234cd5.

The edit mapping was invalid so we just created a dummy GPU
batch for the selection edit data. But it was invalid because we
weren't in edit mode. The fix just requires tweaking the checks
for this data that's also used outside of edit mode.