Fix #132210: Cryptomatte doesn't work with image sequences
The Cryptomatte node doesn't work with image sequences that starts at a
non-zero frame. That's because the typename of the Cryptomatte is not
necessary updated at a frame that exists in the sequence, so the node
fails entirely.

To fix this, we always compute the typename at the frame of the first
image in the sequence.