Fix #132097: Color Tooltips Without Alpha
Currently the popup tooltips for colors shows the same display whether
the color supports alpha or not. This can be confusing when what is
shown could imply that alpha is changeable when it is not.  This PR
makes it clear by not showing alpha values for RBA colors.

Pull Request: https://projects.blender.org/blender/blender/pulls/132287