Tests: Add a render test for a complex scene that can sometimes have missing shadows
While testing the RX 5700XT with HIP on Linux, I noticed that under
specific conditions, shadows would not render properly. The specific
conditions being a combination of:
- Curve rendering
- Subsurface Scattering
- Custom mesh light shaders
- Specific GPU and HIP compiler combinations

This specific combination was not present in the render test suite.
This commit adds a heavily simplified verion of a file that is affected
(Cosmos Laundromat demo) to the test suite to test for this specific
issue.

Ref: blender/blender-test-data!41