Tests: Add render test for distant lights with world volumes
Recently a bug was introduced leading to rendering artifacts on some
devices when rendering a world volume in a scene that also contained a
distant light. This has since been fixed (1).

This commit aims to add a test for this specific case so we can detect
these sorts of issues in the future.

(1) blender/blender@58b7543ede16e2531145272538d909f0fe9ed0f9

Ref: blender/blender-test-data!48