Fix #113685: Knife precision penalized by long-distance constraints
Precision in knife operations is greatly impacted when projecting a 3D
coordinate to 2D in screen space.

This happened for example when performing constraint operations, where
only the mval of the constrained position was used for snapping.

To solve this problem, the constrained position is now used to
calculate the ray_orig and ray_dir used in snap operations.

Now, a ray_orig and ray_dir are calculated whenever an mval is
set and can be adjusted later when setting a constraint.

Ref !131764