Cleanup: Add `NonCopyable` to Paint BVH `Node` class
When working with the Paint BVH, we mostly want to operate on the
existing set of nodes stored in the Tree, even when parititioning and
creating new nodes. Adding `NonCopyable` to the base `Node` class
prevents a subtle class of errors early where a copy is made instead of
acquiring a reference.

Pull Request: https://projects.blender.org/blender/blender/pulls/135547